obs, net: Deduplicate update_nad_references into libs/net/vmspec
Move update_nad_references to libs/net/vmspec.py to eliminate identical
copies in tests/utils.py, tests/network/l2_bridge/nad_ref_change/lib_helpers.py,
and tests/network/libs/nad_ref.py. The function is used across network and
observability team directories, so it belongs in a shared module.
Also adds fail-fast validation: raises ValueError for missing networks,
non-Multus networks, and unmatched interface names.

assisted by: claude code claude-opus-4-6
